+++ src/main/java/com/meme/ala/domain/member/service/MemberCardServiceImpl.java
@@ -65,68 +65,4 @@ public class MemberCardServiceImpl implements MemberCardService {
                 .stream().map(AlaCardSettingPair::getAlaCard)
                 .collect(Collectors.toList());
     }
-
-    @Override
-    @Transactional(readOnly = true)
-    public List<AlaCardDto> getAlaCardDtoList(Member member) {
-        List<AlaCardSettingPair> alaCardSettingPairList = member.getAlaCardSettingPairList();
-        return alaCardSettingPairList.stream()
-                .map(alaCardSettingPair -> alaCardMapper.toDto(
-                        alaCardSettingPair.getAlaCard(),
-                        AlaCardSetting.builder()
-                                .background(Background.builder()
-                                        .fontColor("#ffffff")
-                                        .backgroundColor("#000000")
-                                        .build())
-                                .font(Font.builder()
-                                        .fontName("noto sans")
-                                        .build())
-                                .build(),
-                        this.toSentence(alaCardSettingPair.getAlaCard(), member).getSentence(),
-                        this.toSentence(alaCardSettingPair.getAlaCard(), member).getWordCountList()
-                )).collect(Collectors.toList());
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public SentenceWord toSentence(AlaCard alaCard, Member member) {
-        List<WordCount> wordCountResultList = new ArrayList<>();
-        Aggregation aggregation = aggregationService.findByMember(member);
-        StringBuilder stringBuilder = new StringBuilder();
-        for (int i = 0; i < alaCard.getMiddleCategoryList().size(); i++) {
-            MiddleCategory middleCategory = alaCard.getMiddleCategoryList().get(i);
-            List<WordCount> wordCountList = toSortedWordCountList(aggregation, middleCategory.getMiddleCategoryName());
-            WordCount selectedWordCount = wordCountList.get(0);
-            wordCountResultList.add(selectedWordCount);
-            stringBuilder.append(toMiddleSentence(middleCategory,
-                    member.getMemberSetting().getNickname(),
-                    selectedWordCount.getWord().getWordName(),
-                    i, alaCard.getMiddleCategoryList().size()));
-        }
-        return SentenceWord.builder()
-                .sentence(stringBuilder.toString())
-                .wordCountList(wordCountResultList)
-                .build();
-    }
-
-    private List<WordCount> toSortedWordCountList(Aggregation aggregation, String middleCategoryName) {
-        return aggregation.getWordCountList().stream()
-                .filter(wordCount -> wordCount.getMiddleCategoryName().equals(middleCategoryName))
-                .sorted(Comparator.comparing(WordCount::getCount).reversed())
-                .collect(Collectors.toList());
-    }
-
-    private String toMiddleSentence(MiddleCategory middleCategory, String nickName, String wordName, int idx, int size) {
-        String formPrefix = middleCategory.getPrefix()
-                .replace("???", nickName);
-        String formSentence = middleCategory.getForm()
-                .replace("???", wordName);
-        if (idx == 0) {
-            return formPrefix + formSentence + middleCategory.getSentenceComponent().getJosa();
-        } else if (idx == size - 1) {
-            return formSentence + middleCategory.getSentenceComponent().getEomi();
-        } else {
-            return formSentence + middleCategory.getSentenceComponent().getEomi();
-        }
-    }
 }
